Do not call construction handle in cacheLock
get() method can internally call cache() method inside the saveTaskPool's thread, and the get() method will be blocked until the cache() is done. Previous implementation of cache() also invokes construction handle, and at this point, if construction handle is blocked by cacheLock (which it was, through Observable's update() method), it will wait for the cacheLock to be released, which is blocked by get() method earlier, yet the get() method is also waiting for the cache() method to be done and returned, so they are eventually waiting for each other making a dead lock situation.